#rb none
#rnx

[CL 15275368 by Ben Marsh in ue5-main branch]

+++ Engine/Source/Programs/UnrealGameSync/UnrealGameSync/ArchiveManifest.cs
@@ -1,4 +1,4 @@
-﻿// Copyright Epic Games, Inc. All Rights Reserved.
+// Copyright Epic Games, Inc. All Rights Reserved.
 
 using System;
 using System.Collections.Generic;
@@ -84,25 +84,28 @@ namespace UnrealGameSync
 			DateTime TimeStamp = DateTime.UtcNow;
 			using(ZipFile Zip = new ZipFile(ArchiveFileName))
 			{
-				File.Delete(ManifestFileName);
-
-				// Create the manifest
-				ArchiveManifest Manifest = new ArchiveManifest();
-				foreach(ZipEntry Entry in Zip.Entries)
+				if (ManifestFileName != null)
 				{
-					if(!Entry.FileName.EndsWith("/") && !Entry.FileName.EndsWith("\\"))
-					{ 
-						Manifest.Files.Add(new ArchiveManifestFile(Entry.FileName, Entry.UncompressedSize, TimeStamp));
+					File.Delete(ManifestFileName);
+
+					// Create the manifest
+					ArchiveManifest Manifest = new ArchiveManifest();
+					foreach (ZipEntry Entry in Zip.Entries)
+					{
+						if (!Entry.FileName.EndsWith("/") && !Entry.FileName.EndsWith("\\"))
+						{
+							Manifest.Files.Add(new ArchiveManifestFile(Entry.FileName, Entry.UncompressedSize, TimeStamp));
+						}
 					}
-				}
 
-				// Write it out to a temporary file, then move it into place
-				string TempManifestFileName = ManifestFileName + ".tmp";
-				using(FileStream OutputStream = File.Open(TempManifestFileName, FileMode.Create, FileAccess.Write))
-				{
-					Manifest.Write(OutputStream);
+					// Write it out to a temporary file, then move it into place
+					string TempManifestFileName = ManifestFileName + ".tmp";
+					using (FileStream OutputStream = File.Open(TempManifestFileName, FileMode.Create, FileAccess.Write))
+					{
+						Manifest.Write(OutputStream);
+					}
+					File.Move(TempManifestFileName, ManifestFileName);
 				}
-				File.Move(TempManifestFileName, ManifestFileName);
 
 				// Extract all the files
 				int EntryIdx = 0;
